Began refactoring Stage.scala
