+++ core/src/main/scala/org/apache/spark/scheduler/DAGScheduler.scala
@@ -997,7 +997,7 @@ class DAGScheduler(
               logInfo("running: " + runningStages)
               logInfo("waiting: " + waitingStages)
               logInfo("failed: " + failedStages)
-              if (stage.shuffleDep.isDefined) {
+              if (stage.isShuffleMap) {
                 // We supply true to increment the epoch number here in case this is a
                 // recomputation of the map outputs. In that case, some nodes may have cached
                 // locations with holes (from when we detected the error) and will need the
