+++ core/src/main/scala/org/apache/spark/scheduler/ResultStage.scala
@@ -1,0 +1,8 @@
+package org.apache.spark.scheduler
+
+/**
+ * Created by zjb238 on 2/17/15.
+ */
+class ResultStage {
+
+}
